// src/modules/invoice/enums/invoice-status.enum.ts
export enum INVOICE_STATUS {
  Nonexistent = 'invoice.status.non_existent',
  Draft = 'invoice.status.draft',
  Sent = 'invoice.status.sent',
  Validated = 'invoice.status.validated',
  Paid = 'invoice.status.paid',
  PartiallyPaid = 'invoice.status.partially_paid',
  Unpaid = 'invoice.status.unpaid',
  Expired = 'invoice.status.expired',
  Archived = 'quotation.status.archived',
}